simplify types

+++ src/core/misc.c
@@ -29,7 +29,7 @@
 #include <sys.H>
 #include <types.H>
 
-size_t strlen(const char *X) {
+uint64_t strlen(const char *X) {
 	const char *Y;
 	if (X == NULL) return 0;
 	Y = X;
@@ -64,7 +64,7 @@ int strcmp(const char *X, const char *Y) {
 	return 0;
 }
 
-int strcmpn(const char *X, const char *Y, size_t n) {
+int strcmpn(const char *X, const char *Y, uint64_t n) {
 	while (n > 0 && *X == *Y && *X != '\0' && *Y != '\0') {
 		X++;
 		Y++;
@@ -87,14 +87,14 @@ char *substr(const char *s, const char *sub) {
 	return NULL;
 }
 
-char *substrn(const char *s, const char *sub, size_t n) {
-	size_t i;
+char *substrn(const char *s, const char *sub, uint64_t n) {
+	uint64_t i;
 	if (s == NULL || sub == NULL || n == 0) return NULL;
 	if (*sub == '\0') return (char *)s;
 
 	for (i = 0; i < n && s[i]; i++) {
 		const char *tmps = s + i, *tmpsub = sub;
-		size_t j = i;
+		uint64_t j = i;
 		while (j < n && *tmps == *tmpsub && *tmps) {
 			tmps++;
 			tmpsub++;
@@ -113,9 +113,9 @@ char *strchr(const char *s, int c) {
 	return (*s == c) ? (char *)s : 0;
 }
 
-void *memset(void *dest, int c, size_t n) {
+void *memset(void *dest, int c, uint64_t n) {
 	uint8_t *s = (uint8_t *)dest;
-	size_t i;
+	uint64_t i;
 
 	if (dest == NULL || n == 0) {
 		return dest;
@@ -127,11 +127,11 @@ void *memset(void *dest, int c, size_t n) {
 	return dest;
 }
 
-int memcmp(const void *s1, const void *s2, size_t n) {
+int memcmp(const void *s1, const void *s2, uint64_t n) {
 	/* Cast pointers to unsigned char for byte-by-byte comparison */
 	const unsigned char *p1 = (const unsigned char *)s1;
 	const unsigned char *p2 = (const unsigned char *)s2;
-	size_t i;
+	uint64_t i;
 
 	/* Compare each byte */
 	for (i = 0; i < n; i++) {
@@ -145,10 +145,10 @@ int memcmp(const void *s1, const void *s2, size_t n) {
 	return 0;
 }
 
-void *memcpy(void *dest, const void *src, size_t n) {
+void *memcpy(void *dest, const void *src, uint64_t n) {
 	uint8_t *d = (uint8_t *)dest;
 	const uint8_t *s = (uint8_t *)src;
-	size_t i;
+	uint64_t i;
 
 	if (dest == NULL || src == NULL || n == 0) {
 		return dest;
@@ -160,10 +160,10 @@ void *memcpy(void *dest, const void *src, size_t n) {
 	return dest;
 }
 
-void *memorymove(void *dest, const void *src, size_t n) {
+void *memorymove(void *dest, const void *src, uint64_t n) {
 	uint8_t *d = (uint8_t *)dest;
 	const uint8_t *s = (uint8_t *)src;
-	size_t i;
+	uint64_t i;
 
 	if (dest == NULL || src == NULL || n == 0) {
 		return dest;
@@ -182,9 +182,9 @@ void *memorymove(void *dest, const void *src, size_t n) {
 	return dest;
 }
 
-void byteszero(void *s, size_t len) { memset(s, 0, len); }
+void byteszero(void *s, uint64_t len) { memset(s, 0, len); }
 
-size_t uint128_t_to_string(char *buf, uint128_t v) {
+uint64_t uint128_t_to_string(char *buf, uint128_t v) {
 	char temp[40];
 	int i = 0, j = 0;
 
@@ -206,8 +206,8 @@ size_t uint128_t_to_string(char *buf, uint128_t v) {
 	return j;
 }
 
-size_t int128_t_to_string(char *buf, int128_t v) {
-	size_t len;
+uint64_t int128_t_to_string(char *buf, int128_t v) {
+	uint64_t len;
 	const int128_t int128_min = INT128_MIN;
 	const uint128_t int128_min_abs = (uint128_t)0x8000000000000000UL << 64;
 
@@ -227,9 +227,9 @@ size_t int128_t_to_string(char *buf, int128_t v) {
 }
 
 /* Convert string to unsigned 128-bit integer */
-uint128_t string_to_uint128(const char *buf, size_t len) {
+uint128_t string_to_uint128(const char *buf, uint64_t len) {
 	uint128_t result;
-	size_t i;
+	uint64_t i;
 	char c;
 
 	/* Input validation */
@@ -273,8 +273,8 @@ uint128_t string_to_uint128(const char *buf, size_t len) {
 }
 
 /* Convert string to signed 128-bit integer using string_to_uint128 */
-int128_t string_to_int128(const char *buf, size_t len) {
-	size_t i;
+int128_t string_to_int128(const char *buf, uint64_t len) {
+	uint64_t i;
 	int sign;
 	uint128_t abs_value;
 	int128_t result;
@@ -328,10 +328,10 @@ int128_t string_to_int128(const char *buf, size_t len) {
 	return result;
 }
 
-size_t double_to_string(char *buf, double v, int max_decimals) {
+uint64_t double_to_string(char *buf, double v, int max_decimals) {
 	char temp[41];
-	size_t pos = 0;
-	size_t len;
+	uint64_t pos = 0;
+	uint64_t len;
 	int i;
 	int is_negative;
 	uint64_t int_part;
@@ -424,7 +424,7 @@ size_t double_to_string(char *buf, double v, int max_decimals) {
 	/* Decimal point and fractional part */
 	if (frac_part > 0 && max_decimals > 0) {
 		int digits = 0;
-		size_t frac_start;
+		uint64_t frac_start;
 
 		buf[pos++] = '.';
 		frac_start = pos;
@@ -583,10 +583,10 @@ uint128_t __udivti3(uint128_t a, uint128_t b) {
 }
 int printf(const char *, ...);
 /* Base64 encode */
-size_t b64_encode(const uint8_t *in, size_t in_len, uint8_t *out,
-		  size_t out_max) {
-	size_t i;
-	size_t j;
+uint64_t b64_encode(const uint8_t *in, uint64_t in_len, uint8_t *out,
+		  uint64_t out_max) {
+	uint64_t i;
+	uint64_t j;
 	static const char *b64_table =
 	    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
@@ -630,10 +630,10 @@ size_t b64_encode(const uint8_t *in, size_t in_len, uint8_t *out,
 }
 
 /* Base64 decode */
-size_t b64_decode(const uint8_t *in, size_t in_len, uint8_t *out,
-		  size_t out_max) {
-	size_t i;
-	size_t j;
+uint64_t b64_decode(const uint8_t *in, uint64_t in_len, uint8_t *out,
+		  uint64_t out_max) {
+	uint64_t i;
+	uint64_t j;
 	int b0;
 	int b1;
 	int b2;
@@ -662,7 +662,7 @@ size_t b64_decode(const uint8_t *in, size_t in_len, uint8_t *out,
 
 	/* Account for padding in output size */
 	{
-		size_t expected_out = (in_len / 4) * 3;
+		uint64_t expected_out = (in_len / 4) * 3;
 		if (in_len >= 4 && in[in_len - 1] == '=') out_max++;
 		if (in_len >= 4 && in[in_len - 2] == '=') out_max++;
 		if (out_max < expected_out) {
