+++ src/crypto/aes.c
@@ -232,7 +232,7 @@ static void Cipher(state_t *state, const uint8_t *RoundKey) {
 
 /* Symmetrical operation: same function for encrypting as for decrypting. Note
  * any IV/nonce should never be reused with the same key */
-void aes_ctr_xcrypt_buffer(AesContext *ctx, uint8_t *buf, size_t length) {
+void aes_ctr_xcrypt_buffer(AesContext *ctx, uint8_t *buf, uint64_t length) {
 	uint8_t buffer[AES_BLOCKLEN];
 
 	uint64_t i;
